// Final Evaluation/CabApp/CabAppSPA/src/app/shared/models/Bookings.ts
// Generated by https://quicktype.io

export interface Bookings {
    id:            number;
    email:         string;
    bookingDate:   string;
    bookingTime:   string;
    fromPlace:     number;
    fromPlaceName: string;
    toPlace:       number;
    toPlaceName:   string;
    pickupAddress: string;
    landmark:      string;
    pickupDate:    string;
    pickupTime:    string;
    cabTypeId:     number;
    cabTypeName:   string;
    contactNo:     string;
    status:        string;
}